refactor: rename keydoor.py to key_door.py
