+++ navix/environments/__init__.py
@@ -23,6 +23,7 @@ from __future__ import annotations
 
 from .environment import Environment, Timestep
 from .room import Room
-from .keydoor import KeyDoor
+from .key_door import KeyDoor
 from .four_rooms import FourRooms
 from .key_corridor import KeyCorridor
+from .lava_gap import LavaGap
