+++ navix/environments/key_door.py
@@ -7,9 +7,9 @@ from flax import struct
 from ..components import EMPTY_POCKET_ID
 from ..rendering.cache import RenderingCache
 from ..rendering.registry import PALETTE
-from ..environments import Environment
+from . import Environment
 from ..entities import State, Player, Key, Door, Goal, Wall
-from ..environments import Timestep
+from . import Timestep
 from ..grid import mask_by_coordinates, room, random_positions, random_directions
 from .registry import register_env
